+++ pkg/cache/doc.go
@@ -1,0 +1,5 @@
+package cache
+
+/*
+This package is _heavily_ influenced by go-containerregistry and it's cache implementation: https://github.com/google/go-containerregistry/tree/main/pkg/v1/cache
+*/
