+++ pkg/content/file/file_test.go
@@ -6,13 +6,17 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"os"
-	"path"
 	"path/filepath"
+	"reflect"
 	"testing"
 
 	"github.com/google/go-containerregistry/pkg/name"
+	v1 "github.com/google/go-containerregistry/pkg/v1"
+	"github.com/opencontainers/go-digest"
+	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
 
 	"github.com/rancherfederal/hauler/pkg/apis/hauler.cattle.io/v1alpha1"
+	"github.com/rancherfederal/hauler/pkg/artifact/types"
 	"github.com/rancherfederal/hauler/pkg/content/file"
 	"github.com/rancherfederal/hauler/pkg/log"
 	"github.com/rancherfederal/hauler/pkg/store"
@@ -56,11 +60,11 @@ func TestFile_Copy(t *testing.T) {
 		{
 			name: "should copy a local file successfully without an explicit name",
 			cfg: v1alpha1.File{
-				Ref: f.Name(),
+				Ref:  f.Name(),
+				Name: filepath.Base(f.Name()),
 			},
 			args: args{
 				ctx: ctx,
-				// registry: s.RegistryURL(),
 			},
 		},
 		{
@@ -71,7 +75,6 @@ func TestFile_Copy(t *testing.T) {
 			},
 			args: args{
 				ctx: ctx,
-				// registry: s.RegistryURL(),
 			},
 		},
 		{
@@ -82,7 +85,6 @@ func TestFile_Copy(t *testing.T) {
 			},
 			args: args{
 				ctx: ctx,
-				// registry: s.RegistryURL(),
 			},
 			wantErr: true,
 		},
@@ -93,7 +95,6 @@ func TestFile_Copy(t *testing.T) {
 			},
 			args: args{
 				ctx: ctx,
-				// registry: s.RegistryURL(),
 			},
 		},
 		{
@@ -104,7 +105,6 @@ func TestFile_Copy(t *testing.T) {
 			},
 			args: args{
 				ctx: ctx,
-				// registry: s.RegistryURL(),
 			},
 		},
 	}
@@ -115,17 +115,18 @@ func TestFile_Copy(t *testing.T) {
 				t.Fatal(err)
 			}
 
-			ref, err := name.ParseReference(path.Join("hauler", filepath.Base(tt.cfg.Ref)))
+			ref, err := name.ParseReference(filepath.Base(tt.cfg.Ref))
 			if err != nil {
 				t.Fatal(err)
 			}
 
-			if err := s.Add(ctx, f, ref); (err != nil) != tt.wantErr {
+			_, err = s.Add(ctx, f, ref)
+			if (err != nil) != tt.wantErr {
 				t.Error(err)
 			}
 
-			// if err := f.Copy(tt.args.ctx, tt.args.registry); (err != nil) != tt.wantErr {
-			// 	t.Errorf("Copy() error = %v, wantErr %v", err, tt.wantErr)
+			// if err := validate(tt.cfg.Ref, tt.cfg.Name, m); err != nil {
+			// 	t.Error(err)
 			// }
 		})
 	}
@@ -147,4 +148,41 @@ func (s *testFileServer) Start() *httptest.Server {
 
 func (s *testFileServer) Stop() {
 	s.server.Close()
+}
+
+// validate ensure
+func validate(ref string, name string, got *v1.Manifest) error {
+	data, err := os.ReadFile(ref)
+	if err != nil {
+		return err
+	}
+
+	d := digest.FromBytes(data)
+
+	annotations := make(map[string]string)
+	annotations[ocispec.AnnotationTitle] = name
+	annotations[ocispec.AnnotationSource] = ref
+
+	want := &v1.Manifest{
+		SchemaVersion: 2,
+		MediaType:     types.OCIManifestSchema1,
+		Config:        v1.Descriptor{},
+		Layers: []v1.Descriptor{
+			{
+				MediaType: types.FileLayerMediaType,
+				Size:      int64(len(data)),
+				Digest: v1.Hash{
+					Algorithm: d.Algorithm().String(),
+					Hex:       d.Hex(),
+				},
+				Annotations: annotations,
+			},
+		},
+		Annotations: nil,
+	}
+
+	if !reflect.DeepEqual(want.Layers, got.Layers) {
+		return fmt.Errorf("want = (%v) | got = (%v)", want, got)
+	}
+	return nil
 }
